+++ mteb/abstasks/TaskMetadata.py
@@ -7,7 +7,7 @@ from typing import List, Mapping, Union
 from pydantic import AnyUrl, BaseModel, BeforeValidator, TypeAdapter, field_validator
 from typing_extensions import Annotated, Literal
 
-from .languages import (
+from ..languages import (
     ISO_TO_LANGUAGE,
     ISO_TO_SCRIPT,
     path_to_lang_codes,
