+++ tests/test_TaskMetadata.py
@@ -2,8 +2,8 @@ import logging
 
 import pytest
 
-from mteb.abstasks.TaskMetadata import TaskMetadata
 from mteb.overview import get_tasks
+from mteb.abstasks.TaskMetadata import TaskMetadata
 
 # Historic datasets without filled metadata. Do NOT add new datasets to this list.
 _HISTORIC_DATASETS = [
