+++ tests/test_langscripts.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pytest
 from attr import dataclass
 
-from mteb.abstasks.languages import LanguageScripts
+from mteb.languages import LanguageScripts
 
 
 @dataclass
